refactor(auth): move auth lifecycle and feedback to AuthProvider (#877)
